+++ src/index.ts
@@ -40,7 +40,6 @@ import UserData from "./api/user_data";
 import VideoAttachmentScorer from "./scorer/feature/video_attachment_scorer";
 import { ageInHours, ageInSeconds, ageString, sleep, timeString, toISOFormat } from './helpers/time_helpers';
 import { buildNewFilterSettings, updateHashtagCounts, updateBooleanFilterOptions } from "./filters/feed_filters";
-import { buildTagNames } from './api/objects/tag';
 import { config, MAX_ENDPOINT_RECORDS_TO_PULL, SECONDS_IN_MINUTE } from './config';
 import { FEDIALGO, GIFV, SET_LOADING_STATUS, VIDEO_TYPES, arrowed, extractDomain } from './helpers/string_helpers';
 import { getMoarData, moarDataLogger } from "./api/moar_data_poller";
@@ -719,7 +718,6 @@ export {
     WeightName,
     WeightPresetLabel,
     // Helpers
-    buildTagNames,
     extractDomain,
     isAccessTokenRevokedError,
     isDebugMode,